requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-debug-toolbar==4.2.0
sentry-sdk==1.39.1
django-allauth==0.60.0
google-auth==2.38.0
dj-rest-auth==7.0.1
google-auth-oauthlib==1.2.0
django-filter==24.3
stripe==11.5.0
celery==5.4.0
django-cleanup==9.0.0
paypalrestsdk==1.13.3
cloudinary==1.36.0
django-cloudinary-storage==0.3.0c
